requirements.txt
asgiref==3.7.2
Babel==2.12.1
Django==3.2
django-password-reset==2.0
importlib-metadata==6.8.0
Markdown==3.4.3
mysqlclient==2.2.0
Pillow==7.1.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
zipp==3.16.2
